tools/design-eval/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "design-eval"
version = "0.1.0"
description = "Unified design critique evaluation toolkit"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "datasets>=2.14",
    "Pillow>=10.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "colormath>=3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
full = [
    "torch>=2.0",
    "open-clip-torch>=2.20",
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
    "opencv-python>=4.8",
]
adapters = [
    "openai>=1.0",
    "anthropic>=0.18",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
design-eval = "design_eval.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/design_eval"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
